Feature/RP-75 GDPR Art. 18 | Right to restriction of processing (#705)
* Add a ProcessLocking and unlocking button to the user-page. Aswell added a function with temporally logic

* Add a processingLocked value and getter & setter function

* Add processingLocked to the DataModel

Add processingLocked with the default boolean value of false to the data model

* show the attribute processingLocked within the administration side

* show the attribute processingLocked within the administration side

* Disable buttons in the userview

Disable buttons in the userview if processingLocked is set

* Forbid user updating & deletion if processingLocked is set

* Check ProcessingLocked before updating/deleting a user

* refactor the check of processingLocked for the user update

* Disable buttons if processingLocked is set for user

* Verify if processingLocked is set or not before updating/deleting user

* Add Usermanagement-client to the rolemanagement realm

* Check if the user is locked before making any change of a TenantRoleUser releationship

* Fix issue with checkbox and remove required tag

* Add the new attribute processingLocked to the userFunction.
Refactor functioncalls within classes and tests which call the edited functions

* Replace variable with its value on function calls

* When locking the current user get him from the session

* If no processingLocked value was send insert a false

* refactory the processingLocked function

* Update old tests and add new one for processingLocked

* Update the processingLocked error notification

* Update UserDataModel and its test

* Delete check for the result of saveUser

* Add ErrorMessage

* Fix errors in tests

* Refactor tests and edit logs

* Add test incase processingLocked is set

* Update UserTenantRolesManagerTest.java

* Refactor tests and delete not used function

* Adding testFunction in case processingLocked is set

* Implement Feedback from CodeReview

* Add new class ProcessingLockedException in case user is locked

* Add Missing function isProcessingLocked

* Refactor tests which failed because of the change of how notifications are called

* cleanUp TenantRoleUserBusinessServiceTest

* Refactor tests

* Add new tests

* Add new tests

* add missing function call within test

* Add tests for cases if processingLocked is true

* If the user is locked not buttons for not visible sides should be disabled

* remove the lock from the email change

* Remove comented code

* Within unitTests move the "verify" function to the right position

* If user is locked throw a ProcessingLockedException inside of bussinesServices

* Added missing tests for processingLocked and fix tests

* Fix codesmell

* Add some comments and replace a genericExcpetion

+++ rd-ms-rolemanagement/src/main/java/io/radien/ms/rolemanagement/service/TenantRoleUserBusinessService.java
@@ -25,20 +25,19 @@ import io.radien.api.service.role.exception.RoleException;
 import io.radien.api.service.tenant.ActiveTenantRESTServiceAccess;
 import io.radien.api.service.tenant.TenantRESTServiceAccess;
 import io.radien.api.service.tenantrole.TenantRoleUserServiceAccess;
-import io.radien.exception.BadRequestException;
-import io.radien.exception.GenericErrorCodeMessage;
-import io.radien.exception.InvalidArgumentException;
-import io.radien.exception.SystemException;
+import io.radien.api.service.user.UserRESTServiceAccess;
+import io.radien.exception.*;
 import io.radien.api.service.role.exception.TenantRoleNotFoundException;
 import io.radien.api.service.role.exception.TenantRoleUserNotFoundException;
-import io.radien.exception.UniquenessConstraintException;
 import io.radien.ms.rolemanagement.client.entities.TenantRoleUser;
 import io.radien.ms.rolemanagement.entities.TenantRoleUserEntity;
+
 import java.io.Serializable;
 import java.text.MessageFormat;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
+import java.util .Optional;
 import javax.ejb.Stateless;
 import javax.inject.Inject;
 import javax.ws.rs.core.Response;
@@ -64,6 +63,8 @@ public class TenantRoleUserBusinessService implements Serializable {
     @Inject
     private TenantRESTServiceAccess tenantRESTService;
 
+    @Inject
+    private UserRESTServiceAccess userRESTServiceAccess;
 
     /**
      * Retrieves TenantRoleUser by given id
@@ -122,21 +123,34 @@ public class TenantRoleUserBusinessService implements Serializable {
      * @throws BadRequestException if arguments are invalid or association already exists
      * @throws RoleException in case of duplicate data
      */
+
     public void assignUser(TenantRoleUser tru) {
-        SystemTenantRole tenantRole = tenantRoleService.getById(tru.getTenantRoleId());
-        try {
-            if (tenantRoleUserServiceAccess.isAssociationAlreadyExistent(tru.getUserId(), tru.getTenantRoleId(), tru.getId())) {
-                throw new BadRequestException(
-                        GenericErrorCodeMessage.TENANT_ROLE_USER_IS_ALREADY_ASSOCIATED.toString(
-                                String.valueOf(tenantRole.getTenantId()),
-                                String.valueOf(tenantRole.getRoleId()))
-                );
+        if (!checkIfUserIsLocked(tru.getUserId())) {
+            SystemTenantRole tenantRole = tenantRoleService.getById(tru.getTenantRoleId());
+            try {
+                if (tenantRoleUserServiceAccess.isAssociationAlreadyExistent(tru.getUserId(), tru.getTenantRoleId(), tru.getId())) {
+                    throw new BadRequestException(
+                            GenericErrorCodeMessage.TENANT_ROLE_USER_IS_ALREADY_ASSOCIATED.toString(
+                                    String.valueOf(tenantRole.getTenantId()),
+                                    String.valueOf(tenantRole.getRoleId()))
+                    );
+                }
+                this.tenantRoleUserServiceAccess.create(new TenantRoleUserEntity(tru));
+            } catch (InvalidArgumentException e) {
+                throw new BadRequestException(e.getMessage());
+            } catch (UniquenessConstraintException e) {
+                throw new RoleException(e.getMessage(), Response.Status.BAD_REQUEST);
             }
-            this.tenantRoleUserServiceAccess.create(new TenantRoleUserEntity(tru));
-        } catch (InvalidArgumentException e) {
-            throw new BadRequestException(e.getMessage());
-        } catch (UniquenessConstraintException e) {
-            throw new RoleException(e.getMessage(), Response.Status.BAD_REQUEST);
+        } else {
+            throw new ProcessingLockedException();
+        }
+    }
+
+    protected boolean checkIfUserIsLocked (Long userId){
+        try {
+            return userRESTServiceAccess.isProcessingLocked(userId);
+        } catch (SystemException e) {
+            throw new InternalServerErrorException(e.getMessage());
         }
     }
 
@@ -148,22 +162,27 @@ public class TenantRoleUserBusinessService implements Serializable {
      * @throws RoleException in case of duplicate data
      */
     public void update(Long id, TenantRoleUser tru) {
-        tru.setId(id);
-        SystemTenantRole tenantRole = tenantRoleService.getById(tru.getTenantRoleId());
-        try {
-            if (tenantRoleUserServiceAccess.isAssociationAlreadyExistent(tru.getUserId(), tru.getTenantRoleId(), tru.getId())) {
-                throw new BadRequestException(
-                        GenericErrorCodeMessage.TENANT_ROLE_USER_IS_ALREADY_ASSOCIATED.toString(
-                                String.valueOf(tenantRole.getTenantId()),
-                                String.valueOf(tenantRole.getRoleId()))
-                );
+        if (!checkIfUserIsLocked(tru.getUserId())){
+            tru.setId(id);
+            SystemTenantRole tenantRole = tenantRoleService.getById(tru.getTenantRoleId());
+            try {
+                if (tenantRoleUserServiceAccess.isAssociationAlreadyExistent(tru.getUserId(), tru.getTenantRoleId(), tru.getId())) {
+                    throw new BadRequestException(
+                            GenericErrorCodeMessage.TENANT_ROLE_USER_IS_ALREADY_ASSOCIATED.toString(
+                                    String.valueOf(tenantRole.getTenantId()),
+                                    String.valueOf(tenantRole.getRoleId()))
+                    );
+                }
+                this.tenantRoleUserServiceAccess.update(new TenantRoleUserEntity(tru));
+            } catch (InvalidArgumentException e) {
+                throw new BadRequestException(e.getMessage());
+            } catch (UniquenessConstraintException e) {
+                throw new RoleException(e.getMessage(), Response.Status.BAD_REQUEST);
             }
-            this.tenantRoleUserServiceAccess.update(new TenantRoleUserEntity(tru));
-        } catch (InvalidArgumentException e) {
-            throw new BadRequestException(e.getMessage());
-        } catch (UniquenessConstraintException e) {
-            throw new RoleException(e.getMessage(), Response.Status.BAD_REQUEST);
-        }    }
+        } else {
+            throw new ProcessingLockedException();
+        }
+    }
 
     /**
      * Deletes a Tenant Role Permission association
@@ -174,11 +193,16 @@ public class TenantRoleUserBusinessService implements Serializable {
      */
     public void delete(Long id) {
         SystemTenantRoleUser systemTenantRoleUser = get(id);
-        SystemTenantRole tenantRole = tenantRoleService.getById(systemTenantRoleUser.getTenantRoleId());
-        if(tenantRoleUserServiceAccess.delete(id)) {
-            deleteActiveTenant(systemTenantRoleUser.getUserId(), tenantRole.getTenantId());
+        if (!checkIfUserIsLocked(systemTenantRoleUser.getUserId())){
+
+            SystemTenantRole tenantRole = tenantRoleService.getById(systemTenantRoleUser.getTenantRoleId());
+            if(tenantRoleUserServiceAccess.delete(id)) {
+                deleteActiveTenant(systemTenantRoleUser.getUserId(), tenantRole.getTenantId());
+            } else {
+                throw new TenantRoleUserNotFoundException(MessageFormat.format("No tenant role user found for id {0}", id));
+            }
         } else {
-            throw new TenantRoleUserNotFoundException(MessageFormat.format("No tenant role user found for id {0}", id));
+            throw new ProcessingLockedException();
         }
     }
 
@@ -192,17 +216,22 @@ public class TenantRoleUserBusinessService implements Serializable {
      * @throws RoleException if it was not possible to delete active tenant
      */
     public void unAssignUser(Long tenant, Collection<Long> roles, Long user) {
-        try {
-            Collection<Long> ids = tenantRoleUserServiceAccess.getTenantRoleUserIds(tenant, roles, user);
-            if (ids.isEmpty()) {
-                throw new TenantRoleUserNotFoundException(
-                        TENANT_ROLE_NO_ASSOCIATION_FOUND_FOR_PARAMS.toString(String.valueOf(tenant), String.valueOf(roles), String.valueOf(user))
-                );
+        if (!checkIfUserIsLocked(user)){
+
+            try {
+                Collection<Long> ids = tenantRoleUserServiceAccess.getTenantRoleUserIds(tenant, roles, user);
+                if (ids.isEmpty()) {
+                    throw new TenantRoleUserNotFoundException(
+                            TENANT_ROLE_NO_ASSOCIATION_FOUND_FOR_PARAMS.toString(String.valueOf(tenant), String.valueOf(roles), String.valueOf(user))
+                    );
+                }
+                tenantRoleUserServiceAccess.delete(ids);
+                deleteActiveTenant(user, tenant);
+            } catch (InvalidArgumentException e) {
+                throw new BadRequestException(e.getMessage());
             }
-            tenantRoleUserServiceAccess.delete(ids);
-            deleteActiveTenant(user, tenant);
-        } catch (InvalidArgumentException e) {
-            throw new BadRequestException(e.getMessage());
+        } else {
+            throw new ProcessingLockedException();
         }
     }
